Pull new request

+++ java/debugger/impl/src/com/intellij/debugger/engine/SuspendManagerImpl.java
@@ -1,23 +1,11 @@
-/*
- * Copyright 2000-2016 JetBrains s.r.o.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
+// Copyright 2000-2018 JetBrains s.r.o. Use of this source code is governed by the Apache 2.0 license that can be found in the LICENSE file.
 package com.intellij.debugger.engine;
 
-import com.intellij.debugger.engine.events.DebuggerCommandImpl;
+import com.intellij.debugger.impl.PrioritizedTask;
+import com.intellij.debugger.jdi.JvmtiError;
 import com.intellij.debugger.jdi.ThreadReferenceProxyImpl;
 import com.intellij.openapi.diagnostic.Logger;
+import com.intellij.openapi.util.registry.Registry;
 import com.sun.jdi.InternalException;
 import com.sun.jdi.ObjectCollectedException;
 import com.sun.jdi.event.EventSet;
@@ -26,6 +14,7 @@ import org.intellij.lang.annotations.MagicConstant;
 import org.jetbrains.annotations.NotNull;
 
 import java.util.*;
+import java.util.concurrent.ConcurrentLinkedDeque;
 
 /**
  * @author lex
@@ -33,12 +22,12 @@ import java.util.*;
 public class SuspendManagerImpl implements SuspendManager {
   private static final Logger LOG = Logger.getInstance("#com.intellij.debugger.engine.SuspendManager");
 
-  private final LinkedList<SuspendContextImpl> myEventContexts  = new LinkedList<>();
+  private final Deque<SuspendContextImpl> myEventContexts = new ConcurrentLinkedDeque<>();
   /**
    * contexts, paused at breakpoint or another debugger event requests. Note that thread, explicitly paused by user is not considered as
    * "paused at breakpoint" and JDI prohibits data queries on its stack frames
    */
-  private final LinkedList<SuspendContextImpl> myPausedContexts = new LinkedList<>();
+  private final Deque<SuspendContextImpl> myPausedContexts = new ConcurrentLinkedDeque<>();
   private final Set<ThreadReferenceProxyImpl>  myFrozenThreads  = Collections.synchronizedSet(new HashSet<>());
 
   private final DebugProcessImpl myDebugProcess;
@@ -57,6 +46,7 @@ public class SuspendManagerImpl implements SuspendManager {
     });
   }
 
+  @NotNull
   @Override
   public SuspendContextImpl pushSuspendContext(@MagicConstant(flagsFromClass = EventRequest.class) final int suspendPolicy, int nVotes) {
     SuspendContextImpl suspendContext = new SuspendContextImpl(myDebugProcess, suspendPolicy, nVotes, null) {
@@ -75,8 +65,7 @@ public class SuspendManagerImpl implements SuspendManager {
               catch (InternalException e) {
                 //InternalException 13 means that there are running threads that we are trying to resume
                 //On MacOS it happened that native thread didn't stop while some java thread reached breakpoint
-                //noinspection StatementWithEmptyBody
-                if (/*Patches.MAC_RESUME_VM_HACK && */e.errorCode() == 13) {
+                if (/*Patches.MAC_RESUME_VM_HACK && */e.errorCode() == JvmtiError.THREAD_NOT_SUSPENDED) {
                   //Its funny, but second resume solves the problem
                 }
                 else {
@@ -109,6 +98,7 @@ public class SuspendManagerImpl implements SuspendManager {
     return suspendContext;
   }
 
+  @NotNull
   @Override
   public SuspendContextImpl pushSuspendContext(final EventSet set) {
     SuspendContextImpl suspendContext = new SuspendContextImpl(myDebugProcess, set.suspendPolicy(), set.size(), set) {
@@ -145,8 +135,8 @@ public class SuspendManagerImpl implements SuspendManager {
           catch (InternalException e) {
             //InternalException 13 means that there are running threads that we are trying to resume
             //On MacOS it happened that native thread didn't stop while some java thread reached breakpoint
-            //noinspection StatementWithEmptyBody
-            if (/*Patches.MAC_RESUME_VM_HACK && */e.errorCode() == 13 && set.suspendPolicy() == EventRequest.SUSPEND_ALL) {
+            if (/*Patches.MAC_RESUME_VM_HACK && */e.errorCode() == JvmtiError.THREAD_NOT_SUSPENDED &&
+                                                  set.suspendPolicy() == EventRequest.SUSPEND_ALL) {
               //Its funny, but second resume solves the problem
             }
             else {
@@ -192,7 +182,7 @@ public class SuspendManagerImpl implements SuspendManager {
 
   @Override
   public SuspendContextImpl getPausedContext() {
-    return !myPausedContexts.isEmpty() ? myPausedContexts.getFirst() : null;
+    return myPausedContexts.peekFirst();
   }
 
   public void popContext(SuspendContextImpl suspendContext) {
@@ -221,7 +211,7 @@ public class SuspendManagerImpl implements SuspendManager {
   @Override
   public List<SuspendContextImpl> getEventContexts() {
     DebuggerManagerThreadImpl.assertIsManagerThread();
-    return Collections.unmodifiableList(myEventContexts);
+    return new ArrayList<>(myEventContexts);
   }
 
   @Override
@@ -233,18 +223,13 @@ public class SuspendManagerImpl implements SuspendManager {
   public boolean isSuspended(ThreadReferenceProxyImpl thread) throws ObjectCollectedException{
     DebuggerManagerThreadImpl.assertIsManagerThread();
 
-    boolean suspended = false;
+    boolean suspended;
 
     if (isFrozen(thread)) {
       suspended = true;
     }
     else {
-      for (SuspendContextImpl suspendContext : myEventContexts) {
-        if (suspendContext.suspends(thread)) {
-          suspended = true;
-          break;
-        }
-      }
+      suspended = myEventContexts.stream().anyMatch(suspendContext -> suspendContext.suspends(thread));
     }
 
     //bug in JDI : newly created thread may be resumed even when suspendPolicy == SUSPEND_ALL
@@ -290,7 +275,7 @@ public class SuspendManagerImpl implements SuspendManager {
     }
   }
 
-  private void processVote(final SuspendContextImpl suspendContext) {
+  private void processVote(@NotNull SuspendContextImpl suspendContext) {
     LOG.assertTrue(suspendContext.myVotesToVote > 0);
     suspendContext.myVotesToVote--;
 
@@ -300,30 +285,22 @@ public class SuspendManagerImpl implements SuspendManager {
     if(suspendContext.myVotesToVote == 0) {
       if(suspendContext.myIsVotedForResume) {
         // resume in a separate request to allow other requests be processed (e.g. dependent bpts enable)
-        myDebugProcess.getManagerThread().schedule(new DebuggerCommandImpl() {
-          @Override
-          protected void action() throws Exception {
-            resume(suspendContext);
-          }
-
-          @Override
-          public Priority getPriority() {
-            return Priority.HIGH;
-          }
-        });
+        myDebugProcess.getManagerThread().schedule(PrioritizedTask.Priority.HIGH, () -> resume(suspendContext));
       }
       else {
         LOG.debug("vote paused");
         myDebugProcess.logThreads();
         myDebugProcess.cancelRunToCursorBreakpoint();
-        final ThreadReferenceProxyImpl thread = suspendContext.getThread();
-        myDebugProcess.deleteStepRequests(thread != null ? thread.getThreadReference() : null);
+        if (!Registry.is("debugger.keep.step.requests")) {
+          ThreadReferenceProxyImpl thread = suspendContext.getThread();
+          myDebugProcess.deleteStepRequests(thread != null ? thread.getThreadReference() : null);
+        }
         notifyPaused(suspendContext);
       }
     }
   }
 
-  public void notifyPaused(SuspendContextImpl suspendContext) {
+  public void notifyPaused(@NotNull SuspendContextImpl suspendContext) {
     pushPausedContext(suspendContext);
     myDebugProcess.myDebugProcessDispatcher.getMulticaster().paused(suspendContext);
   }
@@ -341,6 +318,6 @@ public class SuspendManagerImpl implements SuspendManager {
   }
 
   public List<SuspendContextImpl> getPausedContexts() {
-    return myPausedContexts;
+    return new ArrayList<>(myPausedContexts);
   }
 }
