+++ java/java-impl/src/com/intellij/codeInsight/daemon/impl/quickfix/AccessStaticViaInstanceFix.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2000-2014 JetBrains s.r.o.
+ * Copyright 2000-2016 JetBrains s.r.o.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -15,6 +15,7 @@
  */
 package com.intellij.codeInsight.daemon.impl.quickfix;
 
+import com.intellij.codeInsight.BlockUtils;
 import com.intellij.codeInsight.FileModificationService;
 import com.intellij.codeInsight.daemon.QuickFixBundle;
 import com.intellij.codeInsight.daemon.impl.analysis.HighlightMessageUtil;
@@ -22,6 +23,7 @@ import com.intellij.codeInsight.daemon.impl.analysis.HighlightUtil;
 import com.intellij.codeInsight.highlighting.HighlightManager;
 import com.intellij.codeInspection.LocalQuickFixAndIntentionActionOnPsiElement;
 import com.intellij.openapi.application.ApplicationManager;
+import com.intellij.openapi.application.WriteAction;
 import com.intellij.openapi.diagnostic.Logger;
 import com.intellij.openapi.editor.Editor;
 import com.intellij.openapi.editor.colors.EditorColors;
@@ -29,14 +31,12 @@ import com.intellij.openapi.editor.colors.EditorColorsManager;
 import com.intellij.openapi.editor.markup.TextAttributes;
 import com.intellij.openapi.project.Project;
 import com.intellij.psi.*;
-import com.intellij.psi.util.PsiExpressionTrimRenderer;
-import com.intellij.psi.util.PsiTreeUtil;
-import com.intellij.psi.util.PsiUtil;
-import com.intellij.psi.util.PsiUtilCore;
+import com.intellij.psi.util.*;
 import com.intellij.util.IncorrectOperationException;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
+import java.text.MessageFormat;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -62,8 +62,8 @@ public class AccessStaticViaInstanceFix extends LocalQuickFixAndIntentionActionO
     PsiClass aClass = member.getContainingClass();
     if (aClass == null) return "";
     return QuickFixBundle.message("access.static.via.class.reference.text",
-                                  HighlightMessageUtil.getSymbolName(member, substitutor),
-                                  HighlightUtil.formatClass(aClass),
+                                  HighlightMessageUtil.getSymbolName(member, substitutor, PsiFormatUtilBase.SHOW_TYPE),
+                                  HighlightUtil.formatClass(aClass, false),
                                   HighlightUtil.formatClass(aClass, false));
   }
 
@@ -90,24 +90,36 @@ public class AccessStaticViaInstanceFix extends LocalQuickFixAndIntentionActionO
 
     PsiClass containingClass = myMember.getContainingClass();
     if (containingClass == null) return;
-    try {
-      final PsiExpression qualifierExpression = myExpression.getQualifierExpression();
-      PsiElementFactory factory = JavaPsiFacade.getInstance(project).getElementFactory();
-      if (qualifierExpression != null) {
-        if (!checkSideEffects(project, containingClass, qualifierExpression, factory, myExpression,editor)) return;
-        PsiElement newQualifier = qualifierExpression.replace(factory.createReferenceExpression(containingClass));
+    final PsiExpression qualifierExpression = myExpression.getQualifierExpression();
+    PsiElementFactory factory = JavaPsiFacade.getElementFactory(project);
+    if (qualifierExpression != null && !checkSideEffects(project, containingClass, qualifierExpression, factory, myExpression,editor)) return;
+    WriteAction.run(() -> {
+      try {
+        PsiElement newQualifier = factory.createReferenceExpression(containingClass);
+        if (qualifierExpression != null) {
+          newQualifier = qualifierExpression.replace(newQualifier);
+        }
+        else {
+          myExpression.setQualifierExpression((PsiExpression)newQualifier);
+          newQualifier = myExpression.getQualifierExpression();
+        }
         PsiElement qualifiedWithClassName = myExpression.copy();
-        if (myExpression.getTypeParameters().length == 0) {
+        if (myExpression.getTypeParameters().length == 0 && !(containingClass.isInterface() && !containingClass.equals(PsiTreeUtil.getParentOfType(myExpression, PsiClass.class)))) {
           newQualifier.delete();
           if (myExpression.resolve() != myMember) {
             myExpression.replace(qualifiedWithClassName);
           }
         }
       }
-    }
-    catch (IncorrectOperationException e) {
-      LOG.error(e);
-    }
+      catch (IncorrectOperationException e) {
+        LOG.error(e);
+      }
+    });
+  }
+
+  @Override
+  public boolean startInWriteAction() {
+    return false;
   }
 
   private boolean checkSideEffects(final Project project,
@@ -116,7 +128,7 @@ public class AccessStaticViaInstanceFix extends LocalQuickFixAndIntentionActionO
                                    PsiElementFactory factory,
                                    final PsiElement myExpression,
                                    Editor editor) {
-    final List<PsiElement> sideEffects = new ArrayList<PsiElement>();
+    final List<PsiElement> sideEffects = new ArrayList<>();
     boolean hasSideEffects = RemoveUnusedVariableUtil.checkSideEffects(qualifierExpression, null, sideEffects);
     if (hasSideEffects && !myOnTheFly) return false;
     if (!hasSideEffects || ApplicationManager.getApplication().isUnitTestMode()) {
@@ -143,39 +155,30 @@ public class AccessStaticViaInstanceFix extends LocalQuickFixAndIntentionActionO
         @Override
         protected String sideEffectsDescription() {
           if (canCopeWithSideEffects) {
-            return "<html><body>" +
-                   "  There are possible side effects found in expression '" +
-                   qualifierExpression.getText() +
-                   "'<br>" +
-                   "  You can:<ul><li><b>Remove</b> class reference along with whole expressions involved, or</li>" +
-                   "  <li><b>Transform</b> qualified expression into the statement on its own.<br>" +
-                   "  That is,<br>" +
-                   "  <table border=1><tr><td><code>" +
-                   myExpression.getText() +
-                   "</code></td></tr></table><br> becomes: <br>" +
-                   "  <table border=1><tr><td><code>" +
-                   qualifierExpression.getText() +
-                   ";<br>" +
-                   qualifiedWithClassName.getText() +
-                   "       </code></td></tr></table></li>" +
-                   "  </body></html>";
+            return MessageFormat.format(getFormatString(),
+                                        "expression '" + qualifierExpression.getText() + "'",
+                                        myExpression.getText(), //before text
+                                        qualifierExpression.getText() + ";<br>" + qualifiedWithClassName.getText());//after text
           }
           return "<html><body>  There are possible side effects found in expression '" + qualifierExpression.getText() + "'<br>" +
-                 "You can:<ul><li><b>Remove</b> class reference along with whole expressions involved, or</li></body></html>";
+                 "You can <b>Remove</b> class reference along with whole expressions involved</body></html>";
         }
       };
     dialog.show();
     int res = dialog.getExitCode();
     if (res == RemoveUnusedVariableUtil.RemoveMode.CANCEL.ordinal()) return false;
-    try {
-      if (res == RemoveUnusedVariableUtil.RemoveMode.MAKE_STATEMENT.ordinal()) {
-        final PsiStatement statementFromText = factory.createStatementFromText(qualifierExpression.getText() + ";", null);
-        LOG.assertTrue(statement != null);
-        statement.getParent().addBefore(statementFromText, statement);
-      }
-    }
-    catch (IncorrectOperationException e) {
-      LOG.error(e);
+    if (res == RemoveUnusedVariableUtil.RemoveMode.MAKE_STATEMENT.ordinal()) {
+      final PsiStatement statementFromText = factory.createStatementFromText(qualifierExpression.getText() + ";", null);
+      LOG.assertTrue(statement != null);
+      WriteAction.run(() -> {
+        try {
+          PsiElement parent = statement.getParent();
+          BlockUtils.addBefore(parent instanceof PsiForStatement ? (PsiStatement)parent : statement, statementFromText);
+        }
+        catch (IncorrectOperationException e) {
+          LOG.error(e);
+        }
+      });
     }
     return true;
   }
