+++ java/java-impl/src/com/intellij/codeInsight/lookup/VariableLookupItem.java
@@ -1,18 +1,4 @@
-/*
- * Copyright 2000-2015 JetBrains s.r.o.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
+// Copyright 2000-2018 JetBrains s.r.o. Use of this source code is governed by the Apache 2.0 license that can be found in the LICENSE file.
 package com.intellij.codeInsight.lookup;
 
 import com.intellij.codeInsight.AutoPopupController;
@@ -22,7 +8,6 @@ import com.intellij.codeInsight.daemon.impl.JavaColorProvider;
 import com.intellij.codeInsight.daemon.impl.analysis.HighlightControlFlowUtil;
 import com.intellij.codeInsight.lookup.impl.JavaElementLookupRenderer;
 import com.intellij.featureStatistics.FeatureUsageTracker;
-import com.intellij.lang.ASTNode;
 import com.intellij.openapi.editor.Document;
 import com.intellij.openapi.editor.RangeMarker;
 import com.intellij.openapi.util.RecursionManager;
@@ -30,24 +15,23 @@ import com.intellij.openapi.util.registry.Registry;
 import com.intellij.openapi.util.text.StringUtil;
 import com.intellij.psi.*;
 import com.intellij.psi.codeStyle.JavaCodeStyleManager;
-import com.intellij.psi.controlFlow.ControlFlowUtil;
 import com.intellij.psi.impl.source.PostprocessReformattingAspect;
 import com.intellij.psi.impl.source.PsiFieldImpl;
-import com.intellij.psi.impl.source.SourceTreeToPsiMap;
 import com.intellij.psi.util.PsiTreeUtil;
 import com.intellij.psi.util.PsiUtil;
-import com.intellij.util.containers.HashMap;
 import com.intellij.util.ui.ColorIcon;
+import com.intellij.util.ui.JBUI;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
 import java.awt.*;
-import java.util.Collection;
+import java.util.HashMap;
 
 /**
 * @author peter
 */
 public class VariableLookupItem extends LookupItem<PsiVariable> implements TypedLookupItem, StaticallyImportable {
+  private static final String EQ = " = ";
   @Nullable private final MemberLookupHelper myHelper;
   private final Color myColor;
   private final String myTailText;
@@ -80,7 +64,7 @@ public class VariableLookupItem extends LookupItem<PsiVariable> implements Typed
     String initText = initializer == null ? null : initializer.getText();
     if (StringUtil.isEmpty(initText)) return null;
 
-    String prefix = var instanceof PsiEnumConstant ? "" : " = ";
+    String prefix = var instanceof PsiEnumConstant ? "" : EQ;
     String suffix = var instanceof PsiEnumConstant && ((PsiEnumConstant)var).getInitializingClass() != null ? " {...}" : "";
     return StringUtil.trimLog(prefix + initText + suffix, 30);
   }
@@ -157,17 +141,21 @@ public class VariableLookupItem extends LookupItem<PsiVariable> implements Typed
       myHelper.renderElement(presentation, qualify, true, getSubstitutor());
     }
     if (myColor != null) {
-      presentation.setTypeText("", new ColorIcon(12, myColor));
+      presentation.setTypeText("", JBUI.scale(new ColorIcon(12, myColor)));
     } else {
       presentation.setTypeText(getType().getPresentableText());
     }
     if (myTailText != null && StringUtil.isEmpty(presentation.getTailText())) {
-      presentation.setTailText(myTailText, true);
+      if (myTailText.startsWith(EQ)) {
+        presentation.appendTailTextItalic(" (" + myTailText + ")", true);
+      } else {
+        presentation.setTailText(myTailText, true);
+      }
     }
   }
 
   @Override
-  public void handleInsert(InsertionContext context) {
+  public void handleInsert(@NotNull InsertionContext context) {
     PsiVariable variable = getObject();
 
     Document document = context.getDocument();
@@ -187,7 +175,7 @@ public class VariableLookupItem extends LookupItem<PsiVariable> implements Typed
           ref.bindToElementViaStaticImport(((PsiField)variable).getContainingClass());
           PostprocessReformattingAspect.getInstance(ref.getProject()).doPostponedFormatting();
         }
-        if (toDelete.isValid()) {
+        if (toDelete != null && toDelete.isValid()) {
           document.deleteString(toDelete.getStartOffset(), toDelete.getEndOffset());
         }
         context.commitDocument();
@@ -218,7 +206,7 @@ public class VariableLookupItem extends LookupItem<PsiVariable> implements Typed
       TailType.COMMA.processTail(context.getEditor(), context.getTailOffset());
       AutoPopupController.getInstance(context.getProject()).autoPopupParameterInfo(context.getEditor(), null);
     }
-    else if (completionChar == ':' && getAttribute(LookupItem.TAIL_TYPE_ATTR) != TailType.UNKNOWN) {
+    else if (completionChar == ':' && getAttribute(LookupItem.TAIL_TYPE_ATTR) != TailType.UNKNOWN && isTernaryCondition(ref)) {
       context.setAddCompletionChar(false);
       TailType.COND_EXPR_COLON.processTail(context.getEditor(), context.getTailOffset());
     }
@@ -234,6 +222,11 @@ public class VariableLookupItem extends LookupItem<PsiVariable> implements Typed
     }
   }
 
+  private static boolean isTernaryCondition(PsiReferenceExpression ref) {
+    PsiElement parent = ref == null ? null : ref.getParent();
+    return parent instanceof PsiConditionalExpression && ref == ((PsiConditionalExpression)parent).getThenExpression();
+  }
+
   public static void makeFinalIfNeeded(@NotNull InsertionContext context, @NotNull PsiVariable variable) {
     PsiElement place = context.getFile().findElementAt(context.getTailOffset() - 1);
     if (!Registry.is("java.completion.make.outer.variables.final") ||
@@ -242,7 +235,7 @@ public class VariableLookupItem extends LookupItem<PsiVariable> implements Typed
     }
 
     if (HighlightControlFlowUtil.getInnerClassVariableReferencedFrom(variable, place) != null &&
-        !HighlightControlFlowUtil.isReassigned(variable, new HashMap<PsiElement, Collection<ControlFlowUtil.VariableInfo>>())) {
+        !HighlightControlFlowUtil.isReassigned(variable, new HashMap<>())) {
       PsiUtil.setModifierProperty(variable, PsiModifier.FINAL, true);
     }
   }
@@ -271,9 +264,8 @@ public class VariableLookupItem extends LookupItem<PsiVariable> implements Typed
 
     PsiClass containingClass = field.getContainingClass();
     if (containingClass != null && containingClass.getName() != null) {
-      OffsetKey oldStart = context.trackOffset(context.getStartOffset(), true);
+      context.getDocument().insertString(context.getStartOffset(), ".");
       JavaCompletionUtil.insertClassReference(containingClass, file, context.getStartOffset());
-      context.getDocument().insertString(context.getOffsetMap().getOffset(oldStart), ".");
       PsiDocumentManager.getInstance(context.getProject()).commitDocument(context.getDocument());
     }
   }
